requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Flask-CORS==3.0.10
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
dataclasses-json==0.5.14
distro==1.9.0
faiss-cpu==1.8.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.5
langchain-community==0.2.5
# langchain-core==0.2.9
langchain-experimental==0.0.61
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
# langchainplus-sdk==0.0.20
langsmith==0.1.81
lxml==5.2.2
marshmallow==3.21.3
multidict==6.0.5
mypy-extensions==1.0.0
numexpr==2.10.0
numpy==1.26.4
openai==1.35.1
openapi-schema-pydantic==1.2.4
orjson==3.10.5
packaging==24.1
pandas==2.2.2
pydantic==2.7.4
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
tenacity==8.4.1
tiktoken==0.7.0
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
yarl==1.9.4